Move BrukerkontekstConstant to service package
